+++ src/modules/Modules.ts
@@ -1,29 +1,6 @@
 import _ from 'lodash';
 import { flagToEvent, getBiTInfo, lookupFlag } from 'ss-scene-flags';
-import { getOrCreateUserName } from '../util/UserUtils';
-import { getUserByName } from '../database/Users';
-import {
-    Quote,
-    QuoteInfo,
-    addQuote,
-    deleteQuote,
-    editQuote,
-    editQuoteInfo,
-    getLatestQuote,
-    getQuote,
-    getQuoteAlias,
-    getQuoteInfo,
-    getRandomQuote,
-    handleAliasRequest,
-    searchQuote,
-} from '../database/quotes/Quotes';
-import {
-    getCurrency,
-    gambleLoss,
-    gambleWin,
-    addCurrency,
-    getGambleNet,
-} from '../database/Economy';
+import { Quote, QuoteInfo } from '../database/quotes/Quotes';
 
 /**
  * Represents a function that handles a given command or subset of commands
@@ -33,7 +10,7 @@ export type HandlerDelegate = (
     sender: string,
     mod: boolean,
     ...metadata: string[]
-) => Promise<any>;
+) => Promise<string | undefined>;
 
 export type QuoteResultMultiple = {
     quotes: Quote[];
@@ -65,224 +42,6 @@ export type QuoteResult =
     | QuoteResultNotFound
     | QuoteResultMessage
     | QuoteResultError;
-
-const quotePermissionDenied: QuoteResultError = {
-    isPermissionDenied: true,
-    error: 'You do not have permission to do that',
-};
-
-export const handleQuoteCommand: HandlerDelegate = async (
-    commandParts: string[],
-    sender: string,
-    mod: boolean,
-): Promise<QuoteResult> => {
-    const quoteCommand = commandParts[0];
-    if (quoteCommand === 'add') {
-        const quote = commandParts.slice(1).join(' ');
-        const number = addQuote(quote, sender);
-        return {
-            message: `Added quote #${number}`,
-        };
-    }
-    if (quoteCommand === 'delete') {
-        if (!mod) {
-            return quotePermissionDenied;
-        }
-        const quoteNumber = parseInt(commandParts[1], 10);
-        if (!deleteQuote(quoteNumber)) {
-            return {
-                isPermissionDenied: false,
-                error: `Error: ${quoteNumber} is not a number`,
-            };
-        }
-        return {
-            message: `#${quoteNumber} deleted`,
-        };
-    }
-    if (quoteCommand === 'edit') {
-        if (!mod) {
-            return quotePermissionDenied;
-        }
-        const quoteNumber = parseInt(commandParts[1], 10);
-        if (Number.isNaN(quoteNumber)) {
-            return {
-                isPermissionDenied: false,
-                error: `Error: ${quoteNumber} is not a number`,
-            };
-        }
-        const newQuote = commandParts.splice(2).join(' ');
-        editQuote(quoteNumber, newQuote);
-        return {
-            message: `#${quoteNumber} edited`,
-        };
-    }
-    if (quoteCommand === 'alias') {
-        if (!mod) {
-            return quotePermissionDenied;
-        }
-        return {
-            message: handleAliasRequest(commandParts, mod),
-        };
-    }
-    if (quoteCommand === 'info') {
-        if (commandParts[1] === 'edit') {
-            if (!mod) {
-                return quotePermissionDenied;
-            }
-            const quoteNumber = parseInt(commandParts[2], 10);
-            const quotedOn = commandParts[3];
-            const quotedBy = commandParts.slice(4).join(' ');
-            editQuoteInfo(quoteNumber, quotedOn, quotedBy);
-            return {
-                message: `info for #${quoteNumber} updated`,
-            };
-        }
-        const quoteNumber = parseInt(commandParts[1], 10);
-        const results = getQuoteInfo(quoteNumber);
-        if (results === undefined) {
-            return {
-                isPermissionDenied: false,
-                error: 'no quote found',
-            };
-        }
-        return results;
-    }
-    if (quoteCommand === 'search') {
-        const searchString = commandParts.slice(1).join(' ');
-        const results = searchQuote(searchString);
-        if ('error' in results) {
-            return {
-                isPermissionDenied: false,
-                error: results.error,
-            };
-        }
-        if (results.length === 1) {
-            return results[0];
-        }
-        return {
-            quotes: results,
-        };
-    }
-    if (quoteCommand === 'latest') {
-        const quote = getLatestQuote();
-        if (!quote) {
-            return {
-                isPermissionDenied: false,
-                error: 'There are no quotes',
-            };
-        }
-        return quote;
-    }
-    // looking up a quote
-    let quote;
-    if (commandParts.length > 0) {
-        // looking for a specific quote
-        const lookup = commandParts[0];
-        const quoteNumber = parseInt(lookup, 10);
-        if (Number.isNaN(quoteNumber)) {
-            const alias = commandParts.join(' ');
-            quote = getQuoteAlias(alias);
-            if (!quote) {
-                return {
-                    isPermissionDenied: false,
-                    error: `Quote with alias '${alias}' does not exist`,
-                };
-            }
-            return quote;
-        }
-        quote = getQuote(quoteNumber);
-        if (!quote) {
-            return {
-                type: SearchType.number,
-            };
-        }
-    } else {
-        quote = getRandomQuote();
-        if (!quote) {
-            return {
-                isPermissionDenied: false,
-                error: 'There are no quotes',
-            };
-        }
-    }
-    return quote;
-};
-
-export const handleEconomyCommand: HandlerDelegate = async (
-    commandParts: string[],
-    sender: string,
-    mod: boolean,
-    ...metadata: string[]
-): Promise<string> => {
-    const command = commandParts.shift();
-    const [channelName] = metadata;
-    const user = await getOrCreateUserName(sender);
-    const owner = getUserByName(channelName);
-    if (!owner) return 'economy is not configured';
-    const currencyName = 'BiTcoins';
-    if (command === 'money') {
-        let target: string;
-        if (commandParts.length > 0) {
-            [target] = commandParts;
-        } else {
-            target = sender;
-        }
-        return `${getCurrency(
-            await getOrCreateUserName(target),
-            owner.userId,
-        )} ${currencyName}`;
-    }
-    if (command === 'gamble') {
-        const [amount] = commandParts;
-        const total = getCurrency(user, owner.userId);
-        let gambleAmount: number;
-        if (amount === 'all') {
-            gambleAmount = total;
-        } else {
-            gambleAmount = Number(amount);
-            if (Number.isNaN(gambleAmount)) {
-                return 'You must gamble with a number or "all"';
-            }
-            if (gambleAmount < 0) {
-                return "Can't gamble a negative amount";
-            }
-            if (gambleAmount > total) {
-                return "Can't gamble more than you have";
-            }
-        }
-        if (_.random(1) === 0) {
-            gambleLoss(user, owner.userId, gambleAmount);
-            return `You lost ${gambleAmount} ${currencyName}`;
-        }
-        gambleWin(user, owner.userId, gambleAmount);
-        return `You won ${gambleAmount} ${currencyName}`;
-    }
-    if (command === 'give' && mod) {
-        const [receiver, amount] = commandParts;
-        addCurrency(
-            await getOrCreateUserName(receiver),
-            owner.userId,
-            Number(amount),
-        );
-        return `gave ${receiver} ${amount} ${currencyName}`;
-    }
-    if (command === 'net') {
-        let target: string;
-        if (commandParts.length > 0) {
-            [target] = commandParts;
-        } else {
-            target = sender;
-        }
-        const net = getGambleNet(
-            await getOrCreateUserName(target),
-            owner.userId,
-        );
-        return `${target} has net ${net} ${currencyName} from gambling${
-            net < 0 ? '...f' : '...congrats'
-        }`;
-    }
-    return '';
-};
 
 export const handleFlagCommand: HandlerDelegate = async (
     commandParts: string[],
