+++ src/twitch/modules/TwitchModule.ts
@@ -1,0 +1,9 @@
+import { HandlerDelegate } from '../../modules/Modules';
+
+export type TwitchModule = {
+    name: string;
+    key: string;
+    supportsArbitraryCommands?: boolean;
+    arbitraryDelegate?: HandlerDelegate;
+    commandHandlers: Map<string, HandlerDelegate>;
+};
